useAPIPages hook added
